requirements.txt
mlflow==1.5.0
python-dotenv==0.10.3
pandas==0.25.3
scikit-learn==0.22.1
numpy==1.18.1
notebook==6.0.2